+++ apps/core/management/commands/create_build_id.py
@@ -47,7 +47,7 @@ class Command(BaseCommand):
         build_id_file = settings.BASE_DIR / 'build.py'
         build_date_short = datetime.utcnow().strftime('%Y-%m-%d')
 
-        with open(build_id_file, 'w') as file:
+        with open(build_id_file, 'w', encoding='utf-8') as file:
             file.write('BUILD_ID = "%s"\n' % self.build_id)
             file.write('BUILD_ID_SHORT = "%s"\n' % self.build_id[:4])
             file.write('BUILD_DATE_SHORT = "%s"\n' % build_date_short)