+++ apps/core/context_processors.py
@@ -27,7 +27,7 @@ from .utils import to_absolute_url
 
 def base_url(request):
     """Add website base URL and its name to the context."""
-    return {'BASE_URL': to_absolute_url(url=None)}
+    return {'SITE_URL': to_absolute_url(url=None)}
 
 
 def i18n(request):